Moved watson endpoint to routers
